Refactors: 1) parity_wasm_utils and wasmi_utils, 2) use subcommand from main.rs
