--- src/lib.rs
@@ -1,7 +1,0 @@
-extern crate rand;
-extern crate wasmi;
-
-pub mod debug;
-pub mod exec;
-pub mod solver;
-pub mod utils;
